requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
jinja2
python-multipart